:shirt: Refactor into exporter class
